refactor: scanner/utils/containerrootfs->utils/fsutils/
