+++ plugins/runner/internal/runtimehandler/binary/handler.go
@@ -33,7 +33,7 @@ import (
 	"github.com/openclarity/vmclarity/plugins/runner/internal/runtimehandler"
 	"github.com/openclarity/vmclarity/plugins/runner/types"
 	"github.com/openclarity/vmclarity/plugins/sdk-go/plugin"
-	"github.com/openclarity/vmclarity/scanner/utils/containerrootfs"
+	"github.com/openclarity/vmclarity/utils/fsutils/containerrootfs"
 )
 
 type binaryRuntimeHandler struct {
