+++ scanner/go.mod
@@ -26,6 +26,7 @@ require (
 	github.com/openclarity/vmclarity/core v0.7.1
 	github.com/openclarity/vmclarity/plugins/runner v0.7.1
 	github.com/openclarity/vmclarity/plugins/sdk-go v0.7.1
+	github.com/openclarity/vmclarity/utils v0.7.1
 	github.com/openclarity/yara-rule-server v0.3.0
 	github.com/package-url/packageurl-go v0.1.3
 	github.com/samber/slog-logrus/v2 v2.3.0
@@ -481,4 +482,5 @@ replace (
 	github.com/openclarity/vmclarity/core => ../core
 	github.com/openclarity/vmclarity/plugins/runner => ../plugins/runner
 	github.com/openclarity/vmclarity/plugins/sdk-go => ../plugins/sdk-go
+	github.com/openclarity/vmclarity/utils => ../utils
 )
